refactor: added navigation and integration

+++ app/src/main/java/com/health/vita/register/presentation/viewmodel/SignupViewModel.kt
@@ -106,9 +106,9 @@ class SignupViewModel(
                     name = _name.value ?: "",
                     email = _email.value ?: "",
                     //photoUri = _photoUri.value ?: "",
-                    weight = (_weight.value ?: 0f).toDouble(),
+                    weight = (_weight.value ?: 0f),
                     age = _age.value ?: 0,
-                    height = (_height.value ?: 0f).toDouble(),
+                    height = (_height.value ?: 0f),
                     gender = _gender.value ?: "",
                     goal = _goal.value ?: "",
                     activityLevel = _activityLevel.value ?: 0
